turn off retries on some of the test cases
Some tests are checking that a second instantiation fails.
Don't need to bother doing retries on those, as they're supposed to fail.
As such, they'll take the full time for retries and still fail,
which will slow down the testsuite and increase the likelihood
of failures from parallel CPAN smokers.